tests/mle-star-samples/requirements.txt
# Core ML Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep Learning - PyTorch
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# NLP - Transformers and BERT
transformers>=4.20.0
tokenizers>=0.12.0
datasets>=2.0.0
nltk>=3.7

# Gradient Boosting
xgboost>=1.6.0
lightgbm>=3.3.0
catboost>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Experiment Tracking
tensorboard>=2.9.0
mlflow>=1.26.0
wandb>=0.12.0

# Hyperparameter Optimization
optuna>=3.0.0
hyperopt>=0.2.7
ray[tune]>=2.0.0

# Model Interpretation
shap>=0.41.0
lime>=0.2.0.1
eli5>=0.11.0

# Data Processing
numba>=0.56.0
dask>=2022.5.0
modin[all]>=0.15.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.6.0
imageio>=2.19.0

# Jupyter and Development
jupyter>=1.0.0
jupyterlab>=3.4.0
ipywidgets>=7.7.0
tqdm>=4.64.0

# Utilities
pyyaml>=6.0
joblib>=1.1.0
requests>=2.28.0
click>=8.1.0

# Testing
pytest>=7.1.0
pytest-cov>=3.0.0

# Code Quality
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.960

# Web Framework (for deployment demos)
fastapi>=0.78.0
uvicorn>=0.18.0
streamlit>=1.10.0

# Cloud and Deployment
boto3>=1.24.0
azure-storage-blob>=12.11.0
google-cloud-storage>=2.4.0
docker>=5.0.0

# Optional GPU Acceleration
# Uncomment if you have CUDA-capable GPU
# torch>=1.12.0+cu113
# torchvision>=0.13.0+cu113
# torchaudio>=0.12.0+cu113

# Development Dependencies (optional)
# pre-commit>=2.19.0
# sphinx>=5.0.0
# sphinx-rtd-theme>=1.0.0
